python/requirements.txt
numpy==1.26.4
opencv-python==4.10.0.82
pyFirmata==1.1.0
pyserial==3.5
